process-image/requirements.txt
# Function dependencies, for example:
# package>=version
opencv-python==4.2.0.34
numpy==1.18.2
google-cloud-storage==1.27.0